Update category and merchant management

+++ seller/delete_category.php
@@ -1,11 +1,12 @@
 <?php
-include("../connection/connect.php");
-error_reporting(0);
-session_start();
+session_start(); // temp session
+error_reporting(0); // hide undefined index errors
+include("./../connection/connect.php"); // connection to database
 
-
-// sending query
-mysqli_query($db,"DELETE FROM res_category WHERE c_id = '".$_GET['cat_del']."'");
-header("location:add_category.php");  
-
+if (isset($_SESSION["adm_co"]) && ($_SESSION["adm_co"] == "SUPA"))
+{
+    // sending query
+    mysqli_query($db,"DELETE FROM res_category WHERE c_id = '".$_GET['cat_del']."'");
+    header("location:add_category.php");  
+}
 ?>
